requirements.txt
python-dotenv
azure-search-documents==11.6.0b1
azure-identity
azure-keyvault-secrets
ipykernel
pandas
pyodbc